+++ python/rateslib/curves/curves.py
@@ -24,14 +24,15 @@ from rateslib.calendars import CalInput, add_tenor, dcf
 from rateslib.calendars.dcfs import _DCF1d
 from rateslib.calendars.rs import CalTypes, get_calendar
 from rateslib.default import NoInput, PlotOutput, _drb, _validate_states, _WithState, plot
-from rateslib.dual import (  # type: ignore[attr-defined]
+from rateslib.dual import (
     Dual,
     Dual2,
     dual_exp,
     dual_log,
     set_order_convert,
 )
-from rateslib.dual.utils import Arr1dF64, Arr1dObj, DualTypes, Number
+from rateslib.dual.utils import DualTypes, Number
+from rateslib.dual.variable import Arr1dF64, Arr1dObj
 from rateslib.rs import Modifier, index_left_f64
 from rateslib.rs import from_json as from_json_rs
 from rateslib.splines import PPSplineDual, PPSplineDual2, PPSplineF64
